Refactor core-info-form-step using core component structure
